backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
channels_redis
channels
daphne
deepdiff
dj-database-url
psycopg2-binary